feat: implement mobile checkout steps and order summary
- Add MobileOrderSummary component for displaying order details on mobile.
- Create StepContact component for collecting user contact information.
- Implement StepDelivery component for selecting delivery methods.
- Add StepReview component for reviewing order details and applying promo codes.
- Create StepShipping component for entering shipping address.
- Introduce useCheckout hook to manage checkout state and logic.
